requirements.txt
os
python-dotenv
pandas
web3
python-bittrex
